Add `template_body` to configuration
This enables directly specifying cloudformation templates within the
stack configuration.

This removes the need for the user to have a working S3 bucket available
with their templates. It's not ideal long term, but it's nicer for new
developers to start here and then move to more advanced management by
creating a repository of re-usable templates.